Refactor post initialization and add event-driven structure
Refactored `initArticle` and `initFeaturedPosts` to improve modularity and moved initialization logic to an event-driven system using CustomEvents. Adjusted HTML class names for consistency and added unit tests to verify post initialization workflows. This enhances maintainability and expands test coverage.
