+++ pages/featured-posts-sidebar.html
@@ -6,7 +6,7 @@
     <h2>Featured Posts</h2>
     <ul class="article-list">
       <!-- Article 1 -->
-      <li><article class="article">
+      <li><article class="post-summary">
         <i class=""></i>
         <h3>Article Title</h3>
         <p>Phasellus quam turpis, feugiat sit amet ornare in, a
@@ -16,7 +16,7 @@
       </article></li>
 
       <!-- Article 2 -->
-      <li><article class="article">
+      <li><article class="post-summary">
         <i class=""></i>
         <h3>Article Title</h3>
         <p>Phasellus quam turpis, feugiat sit amet ornare in, a
@@ -26,7 +26,7 @@
       </article></li>
 
       <!-- Article 3 -->
-      <li><article class="article">
+      <li><article class="post-summary">
         <i class=""></i>
         <h3>Article Title</h3>
         <p>Phasellus quam turpis, feugiat sit amet ornare in, a
@@ -36,7 +36,7 @@
       </article></li>
 
       <!-- Article 4 -->
-      <li><article class="article">
+      <li><article class="post-summary">
         <i class=""></i>
         <h3>Article Title</h3>
         <p>Phasellus quam turpis, feugiat sit amet ornare in, a
@@ -62,7 +62,7 @@
     padding-top: 1rem;
   }
 
-  .featured-posts-sidebar .article {
+  .featured-posts-sidebar .post-summary {
     margin: 2rem 0;
   }
