+++ pages/featured-posts-small.html
@@ -6,7 +6,7 @@
     <div class="major"><h2>Featured Posts</h2></div>
     <div class="row gtr-50">
       <!-- Article 1 -->
-      <article class="col-3 col-12-narrower article">
+      <article class="col-3 col-12-narrower post-summary">
         <i class=""></i>
         <h3>Article Title</h3>
         <p>Phasellus quam turpis, feugiat sit amet ornare in, a
@@ -16,7 +16,7 @@
       </article>
 
       <!-- Article 2 -->
-      <article class="col-3 col-12-narrower article">
+      <article class="col-3 col-12-narrower post-summary">
         <i class=""></i>
         <h3>Article Title</h3>
         <p>Phasellus quam turpis, feugiat sit amet ornare in, a
@@ -26,7 +26,7 @@
       </article>
 
       <!-- Article 3 -->
-      <article class="col-3 col-12-narrower article">
+      <article class="col-3 col-12-narrower post-summary">
         <i class=""></i>
         <h3>Article Title</h3>
         <p>Phasellus quam turpis, feugiat sit amet ornare in, a
@@ -36,7 +36,7 @@
       </article>
 
       <!-- Article 4 -->
-      <article class="col-3 col-12-narrower article">
+      <article class="col-3 col-12-narrower post-summary">
         <i class=""></i>
         <h3>Article Title</h3>
         <p>Phasellus quam turpis, feugiat sit amet ornare in, a
